pyproject.toml
[tool.poetry]
name = "lego-schedule"
version = "21.2.0"
description = "Auto scheduling for lego small group."
authors = ["Kip Yin <28321392+kipyin@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^20.3.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
